requirements.txt
numpy
tensorflow
pillow
opencv-python
streamlit
